Tweak for making use of python2.7
The "nfs-ganesha.spec-in.cmake" file has OS and OS version checks, which
helps in deciding which python version should be used. For RHEL 7.x python2.7
is the default version, but due to wrong if-else statement for SUSE, for
RHEL 7.x version python3-dbus rpm becomes install time requirement. Hence the
code change.
The "scripts/ganeshactl/CMakeLists.txt" file checks for python3 program (for
being compatible on RHEL 8.x). EOS-core has dependancy on python3, and hence
it is available on the build system. Due to that python3 gets picked up
which conflicts with default python version for RHEL 7.x (python2.7). By
this code change, currently forcing to just look for python2.7 only.

Signed-off-by: Sachin Punadikar <sachin.punadikar@seagate.com>